dev_requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==23.1.0
automium==0.2.6
billiard==4.1.0
celery==5.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
cron-descriptor==1.4.0
Django==4.1.3
django-celery-beat==2.5.0
django-redis==5.2.0
django-timezone-field==5.0
djangorestframework==3.14.0
Flask==2.2.2
flower==1.2.0
future==0.18.3
humanize==4.6.0
idna==3.4
importlib-metadata==5.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.3.0
legacy==0.1.6
line-bot-sdk==2.4.2
MarkupSafe==2.1.1
multidict==6.0.4
mysqlclient==2.1.1
Pillow==9.4.0
prometheus-client==0.17.0
prompt-toolkit==3.0.38
python-crontab==2.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
redis==4.5.5
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
tornado==6.3.2
typing_extensions==4.6.3
tzdata==2022.7
urllib3==1.26.16
vine==5.0.0
wcwidth==0.2.6
Werkzeug==2.2.2
yarl==1.9.2
zipp==3.11.0
